GAE/Tasks/SlightlyNotSoSimpleTask/CloudFunctionCode/task_creator/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
google-cloud-tasks==1.1.0
googleapis-common-protos==1.6.0
wheel
